+++ apps/community/src/entities/articles/index.ts
@@ -1,3 +1,1 @@
-export { getAllArticles, getArticles } from "./action";
-export type { Article } from "./model";
 export { ArticleDetail } from "./ui/ArticleDetail";
