+++ apps/community/src/entities/question/index.ts
@@ -1,5 +1,4 @@
 // Public API
 export { QuestionDetail } from "./ui/QuestionDetail";
-export { QuestionDetailSkeleton } from "./ui/QuestionDetailSkeleton";
 export { QuestionsHeader } from "./ui/QuestionsHeader";
 export { QuestionsList } from "./ui/QuestionsList";
