--- apps/community/src/entities/question/ui/QuestionDetailSkeleton.tsx
@@ -1,27 +1,0 @@
-export function QuestionDetailSkeleton() {
-  return (
-    <div className="animate-pulse space-y-4">
-      <div className="flex items-center justify-between">
-        <div className="flex items-center gap-2">
-          <div className="flex h-8 w-8 items-center justify-center rounded-full bg-gray-200" />
-          <div className="flex flex-col">
-            <div className="h-4 w-20 rounded bg-gray-200" />
-            <div className="mt-1 h-3 w-16 rounded bg-gray-200" />
-          </div>
-        </div>
-        <div className="flex items-center gap-2">
-          <div className="h-8 w-16 rounded bg-gray-200" />
-        </div>
-      </div>
-      <div>
-        <div className="mb-4 h-8 w-3/4 rounded bg-gray-200" />
-        <div className="space-y-2">
-          <div className="h-4 w-full rounded bg-gray-200" />
-          <div className="h-4 w-full rounded bg-gray-200" />
-          <div className="h-4 w-5/6 rounded bg-gray-200" />
-          <div className="h-4 w-3/4 rounded bg-gray-200" />
-        </div>
-      </div>
-    </div>
-  );
-}
