+++ apps/community/src/features/comment/ui/index.ts
@@ -1,6 +1,2 @@
-export { CommentDeleteButton } from "./CommentDeleteButton";
 export { CommentForm } from "./CommentForm";
-export { CommentItem } from "./CommentItem";
-export { CommentLikeButton } from "./CommentLikeButton";
 export { CommentList } from "./CommentList";
-export { CommentReplyButton } from "./CommentReplyButton";
